Remove vertex-edge-intersection check in FrontUpdate.h
The call to `vertex.intersects_mesh_edges()`, which checked if a
vertex is placed too close to its surrounding edges has been removed.
This check led to wrong element sizes when choosing quad layers with
very small heights (compared to the global element size).
It was also found that this check is not required to pass the
current meshing tests and thus has been removed.